% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticutils.R
\name{find_internal_symbols}
\alias{find_internal_symbols}
\title{Given an environment, find all internal symbols used by each object}
\usage{
find_internal_symbols(env = getNamespace("staticutils"))
}
\arguments{
\item{env}{An environment; typically, a namespace for a package.}
}
\value{
A named list, where the name of each element is the name of each
object in \code{env}, and the value is a character vector of strings, where each
string represents a symbol of an object in the environment.
}
\description{
Given an environment, find all internal symbols used by each object
}
\examples{
# By default, find symbols internal to the staticutils package
find_internal_symbols()

# Find all symbols internal to the utils package
find_internal_symbols(getNamespace("utils"))

}
