% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticutils.R
\name{find_internal_symbols_recursive}
\alias{find_internal_symbols_recursive}
\title{Given an environment, find all internal dependencies used by each object,
recursively}
\usage{
find_internal_symbols_recursive(env = getNamespace("staticutils"))
}
\arguments{
\item{env}{An environment; typically, a namespace for a package.}
}
\description{
Given an environment, find all internal dependencies used by each object,
recursively
}
