% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticimports.R
\name{find_symbols}
\alias{find_symbols}
\title{Find all symbols used in a function or language object.}
\usage{
find_symbols(x)
}
\arguments{
\item{x}{A language object or function.}
}
\value{
A character vector of the symbols used in the object.
}
\description{
Find all symbols used in a function or language object.
}
\examples{
# Find all symbols used in the lm function
find_symbols(utils::install.packages)

# Does not find names of arguments
find_symbols(quote(foo(a = A)))
#> [1] "A"   "foo"

find_symbols(quote(function(x = X) NULL))
#> [1] "function" "X"

}
