% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticimports-pkg.R
\name{import_objs}
\alias{import_objs}
\title{Statically import specific objects}
\usage{
import_objs(
  names,
  outfile = here::here("R/staticimports.R"),
  env = "staticimports"
)
}
\arguments{
\item{names}{A character vector of names of objects to import.}

\item{outfile}{File to write to. Defaults to R/staticimports.R in the
current project. Use \code{stdout()} to output to console.}

\item{env}{A string naming a package, or an environment.}
}
\description{
Statically import specific objects
}
\examples{
if (interactive()) {
# Import `os_name` and `walk` into your project
import_objs(c("os_name", "walk"))
}

# Write to stdout instead of R/staticimports.R
import_objs(c("os_name", "walk"), stdout())
}
