% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{staticexport_roclet}
\alias{staticexport_roclet}
\title{staticexport roclet}
\usage{
staticexport_roclet()
}
\value{
A \pkg{roxygen2} roclet
}
\description{
Find all functions tagged with \verb{@staticexport} in function documentation
and copy them into \verb{inst/staticexports/} when documentation is built.

The value of the \verb{@staticexports} tag determines the file the function will
be copied into.
For example, a function with

\if{html}{\out{<div class="sourceCode">}}\preformatted{#' @staticexport foo
}\if{html}{\out{</div>}}

would be copied into \code{inst/staticexports/foo.R}

If the tag appears with no value, e.g.

\if{html}{\out{<div class="sourceCode">}}\preformatted{#' @staticexport
}\if{html}{\out{</div>}}

the file in \verb{inst/staticexports/} will have the same name as the file in \verb{R/}
containing the function.
}
\details{
Note that this function should never need to be called directly.
It only exists to let \verb{@staticexport} tags
be parsed from \pkg{roxygen2} documentation.
}
\examples{
staticexport_roclet()
}
