% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{find_staticimports}
\alias{find_staticimports}
\title{Find any static imports in the comments of the R/ directory}
\usage{
find_staticimports(dir = here::here("R/"))
}
\arguments{
\item{dir}{A directory to look in. Defaults to}
}
\value{
A list of lists. Each sub-list has a \code{source} field and \code{names}.
}
\description{
This finds staticimports declarations, which are comment blocks that look
like:
}
\details{
\preformatted{# @staticimports pkg:staticimports
#   map map_chr map_lgl os_name any_named any_unnamed
#   all_named all_unnamed
}
}
