% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticimports-pkg.R
\name{import}
\alias{import}
\title{Statically import objects}
\usage{
import(
  dir = here::here("R/"),
  outfile = here::here("R/staticimports.R"),
  env = "staticimports"
)
}
\arguments{
\item{dir}{A directory}

\item{outfile}{File to write to. Defaults to R/staticimports.R in the
current project. Use \code{stdout()} to output to console.}

\item{env}{A string naming a package, or an environment.}
}
\description{
This finds staticimports declarations, which are comment blocks that look
like this:\preformatted{# @staticimports map map_chr map_lgl os_name any_named any_unnamed
# all_named all_unnamed
}

It statically imports the named objects into a file, by default
\code{R/staticimports.R}.
}
