% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Statically import one or more functions}
\usage{
import(names, outfile = here::here("R/staticimports.R"), env = "staticimports")
}
\arguments{
\item{names}{A character vector of names of objects to import.}

\item{outfile}{File to write to. Defaults to R/staticimports.R in the
current project. Use \code{stdout()} to output to console.}

\item{env}{A string naming a package, or an environment.}
}
\description{
Statically import one or more functions
}
\examples{
if (interactive()) {
# Import `os_name` and `walk` into your project
import(c("os_name", "walk"))
}

# Write to stdout instead of R/staticimports.R
import(c("os_name", "walk"), stdout())
}
