% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_objs}
\alias{import_objs}
\title{Statically import specific objects}
\usage{
import_objs(
  names,
  source = system.file("staticexports", package = "staticimports"),
  outfile = here::here("R/staticimports.R"),
  label = NULL,
  append = FALSE
)
}
\arguments{
\item{names}{A character vector of names of objects to import.}

\item{source}{A directory containing source files, or an environment to use
as the source.}

\item{outfile}{File to write to. Defaults to R/staticimports.R in the current
project. Use \code{stdout()} to output to console.}

\item{label}{A label to write to the file, to indicate where the objects were
imported from.}

\item{append}{If \code{TRUE}, append to the output file; otherwise overwrite.}
}
\description{
Statically import specific objects
}
\examples{
if (interactive()) {
# Import `os_name` and `walk` into your project
import_objs(c("os_name", "walk"))
}

# Write to stdout instead of R/staticimports.R
import_objs(c("os_name", "walk"), outfile = stdout())
}
